// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/OfferResponse.ts
export interface OfferResponse {
    /**
     * 
     * @type {string}
     * @memberof OfferResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof OfferResponse
     */
    organisationId?: string;
    /**
     * 
     * @type {string}
     * @memberof OfferResponse
     */
    offerType?: OfferResponseOfferTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof OfferResponse
     */
    offerStatus?: OfferResponseOfferStatusEnum;
    /**
     * 
     * @type {OfferDetails}
     * @memberof OfferResponse
     */
    offerDetails?: OfferDetails;
    /**
     * 
     * @type {Date}
     * @memberof OfferResponse
     */
    creationDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof OfferResponse
     */
    expirationDate?: Date;
    /**
     * 
     * @type {RepaymentParametersResponse}
     * @memberof OfferResponse
     */
    basicRepaymentDetails?: RepaymentParametersResponse;
    /**
     * 
     * @type {string}
     * @memberof OfferResponse
     */
    dealId?: string;
}


/**
 * @export
 */
export const OfferResponseOfferTypeEnum = {
    Flat: 'FLAT',
    Fixed: 'FIXED',
    FixedCustomizable: 'FIXED_CUSTOMIZABLE',
    AgreementRestructure: 'AGREEMENT_RESTRUCTURE',
    LineOfCredit: 'LINE_OF_CREDIT',
    InterestRateLineOfCredit: 'INTEREST_RATE_LINE_OF_CREDIT',
    Rbf: 'RBF',
    DailyPayout: 'DAILY_PAYOUT',
    Unknown: 'UNKNOWN'
} as const;
export type OfferResponseOfferTypeEnum = typeof OfferResponseOfferTypeEnum[keyof typeof OfferResponseOfferTypeEnum];

/**
 * @export
 */
export const OfferResponseOfferStatusEnum = {
    New: 'NEW',
    Selected: 'SELECTED',
    PreSigned: 'PRE_SIGNED',
    Signed: 'SIGNED',
    Activated: 'ACTIVATED',
    Closed: 'CLOSED',
    Expired: 'EXPIRED',
    Unknown: 'UNKNOWN'
} as const;
export type OfferResponseOfferStatusEnum = typeof OfferResponseOfferStatusEnum[keyof typeof OfferResponseOfferStatusEnum];